Database setup - Post and Put functions modified
- The database was set using the file driverConfig.go
- Post and Put functions were modified to support the database (both on models files and shoe.go functions)
- main.go modified to initialize and use the database

+++ handler/shoe_put.go
@@ -22,7 +22,7 @@ func ShoeUpdate(store shoe.Updater) gin.HandlerFunc {
 			Style:    request.Style,
 			Color:    request.Color,
 			Material: request.Material,
-			Price:    float32(price_converted),
+			Price:    price_converted,
 		}
 
 		store.UpdateShoe(shoe)
